backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.30
alembic>=1.13.1
pyodbc>=5.1.0
pydantic>=2.7.1
pydantic-settings>=2.2.1
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
redis>=5.0.4
python-dotenv>=1.0.1
python-multipart>=0.0.9
httpx>=0.27.0
pytest>=8.2.0
pytest-asyncio>=0.23.7
